requirements.txt
cffi==1.14.5
colorama==0.4.4
cryptography==3.4.7
falcon==3.0.0
greenlet==1.0.0
gunicorn==20.1.0
jwcrypto==0.8
psycopg2-binary==2.8.6
pycparser==2.20
python-decouple==3.4
simplejson==3.17.2
SQLAlchemy==1.4.8
sqlparse==0.4.1
structlog==21.1.0
tabulate==0.8.9
yoyo-migrations==7.3.1
sqlalchemy-mixins==1.3
prometheus-client==0.11.0
consulate==0.6.0
